ADDED : Token and in controller some changes

+++ src/main/java/com/bridgelabz/lmsmentorservice/service/MentorsService.java
@@ -5,9 +5,11 @@ import com.bridgelabz.lmsmentorservice.dto.MentorsDTO;
 import com.bridgelabz.lmsmentorservice.exception.LMSException;
 import com.bridgelabz.lmsmentorservice.model.MentorsModel;
 import com.bridgelabz.lmsmentorservice.repository.MentorsRepository;
+import com.bridgelabz.lmsmentorservice.util.Response;
 import com.bridgelabz.lmsmentorservice.util.TokenUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.RestTemplate;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -26,6 +28,8 @@ public class MentorsService implements IMentorsService {
     MailService mailService;
     @Autowired
     MentorsRepository mentorsRepository;
+    @Autowired
+    RestTemplate restTemplate;
 
     /*
      * Purpose : Implement the Logic of Creating Mentor Details
@@ -33,14 +37,19 @@ public class MentorsService implements IMentorsService {
      * @Param :  token and mentorsDTO
      * */
     @Override
-    public MentorsModel addMentor(MentorsDTO mentorsDTO) {
-        MentorsModel mentorsModel = new MentorsModel(mentorsDTO);
-        mentorsModel.setCreatedTimeStamp(LocalDateTime.now());
-        mentorsRepository.save(mentorsModel);
-        String body = "Mentor Added Successfully With Id is : " + mentorsModel.getId();
-        String subject = "Mentor Registration Successfully ...";
-        mailService.send(mentorsModel.getEmail(), body, subject);
-        return mentorsModel;
+    public Response addMentor(String token, MentorsDTO mentorsDTO) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            MentorsModel mentorsModel = new MentorsModel(mentorsDTO);
+            mentorsModel.setCreatedTimeStamp(LocalDateTime.now());
+            mentorsRepository.save(mentorsModel);
+            String body = "Mentor Added Successfully With Id is : " + mentorsModel.getId();
+            String subject = "Mentor Registration Successfully ...";
+            mailService.send(mentorsModel.getEmail(), body, subject);
+            return new Response(200, "Success", mentorsModel);
+        }
+        throw new LMSException(400, "Token Wrong");
+
     }
 
     /*
@@ -49,18 +58,23 @@ public class MentorsService implements IMentorsService {
      * @Param :  token,id and profilePic
      * */
     @Override
-    public MentorsModel addProfilePic(Long id, String profilePic) {
-        Optional<MentorsModel> isMentorPresent = mentorsRepository.findById(id);
-        if (isMentorPresent.isPresent()) {
-            isMentorPresent.get().setProfilePic(profilePic);
-            mentorsRepository.save(isMentorPresent.get());
-            String body = "Mentor profilePic Added With Id is : " + isMentorPresent.get().getId();
-            String subject = "Mentor ProfilePic Uploaded ...";
-            mailService.send(isMentorPresent.get().getEmail(), body, subject);
-            return isMentorPresent.get();
-        } else {
-            throw new LMSException(400, "Not found with this id");
+    public Response addProfilePic(String token, Long id, String profilePic) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Optional<MentorsModel> isMentorPresent = mentorsRepository.findById(id);
+            if (isMentorPresent.isPresent()) {
+                isMentorPresent.get().setProfilePic(profilePic);
+                mentorsRepository.save(isMentorPresent.get());
+                String body = "Mentor profilePic Added With Id is : " + isMentorPresent.get().getId();
+                String subject = "Mentor ProfilePic Uploaded ...";
+                mailService.send(isMentorPresent.get().getEmail(), body, subject);
+                return new Response(200, "Success", isMentorPresent.get());
+            } else {
+                throw new LMSException(400, "Not found with this id");
+            }
         }
+        throw new LMSException(400, "Token Wrong");
+
     }
 
     /*
@@ -69,13 +83,17 @@ public class MentorsService implements IMentorsService {
      * @Param :  token
      * */
     @Override
-    public List<MentorsModel> getAllMentors() {
-        List<MentorsModel> isMentors = mentorsRepository.findAll();
-        if (isMentors.size() > 0) {
-            return isMentors;
-        } else {
-            throw new LMSException(400, "Mentors not found");
+    public List<MentorsModel> getAllMentors(String token) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            List<MentorsModel> isMentors = mentorsRepository.findAll();
+            if (isMentors.size() > 0) {
+                return isMentors;
+            } else {
+                throw new LMSException(400, "Mentors not found");
+            }
         }
+        throw new LMSException(400, "Token Wrong");
     }
 
     /*
@@ -84,32 +102,36 @@ public class MentorsService implements IMentorsService {
      * @Param :  token,id and mentorsDTO
      * */
     @Override
-    public MentorsModel updateMentorDetails(Long id, MentorsDTO mentorsDTO) {
-        Optional<MentorsModel> isMentorPresent = mentorsRepository.findById(id);
-        if (isMentorPresent.isPresent()) {
-            isMentorPresent.get().setEmployeeId(mentorsDTO.getEmployeeId());
-            isMentorPresent.get().setFirstName(mentorsDTO.getFirstName());
-            isMentorPresent.get().setLastName(mentorsDTO.getLastName());
-            isMentorPresent.get().setMentorDesc(mentorsDTO.getMentorDesc());
-            isMentorPresent.get().setMentorRole(mentorsDTO.getMentorRole());
-            isMentorPresent.get().setMentorsType(mentorsDTO.getMentorsType());
-            isMentorPresent.get().setEmail(mentorsDTO.getEmail());
-            isMentorPresent.get().setStatus(mentorsDTO.getStatus());
-            isMentorPresent.get().setExperienceYears(mentorsDTO.getExperienceYears());
-            isMentorPresent.get().setMobileNumber(mentorsDTO.getMobileNumber());
-            isMentorPresent.get().setStartDate(mentorsDTO.getStartDate());
-            isMentorPresent.get().setPreferredTime(mentorsDTO.getPreferredTime());
-            isMentorPresent.get().setCreatorUser(mentorsDTO.getCreatorUser());
-            isMentorPresent.get().setSupervisorId(mentorsDTO.getSupervisorId());
-            isMentorPresent.get().setUpdatedTimeStamp(LocalDateTime.now());
-            mentorsRepository.save(isMentorPresent.get());
-            String body = "Mentors Details Updated With Id is : " + isMentorPresent.get().getId();
-            String subject = "Mentors Details Updated Successfully ...";
-            mailService.send(isMentorPresent.get().getEmail(), body, subject);
-        } else {
-            throw new LMSException(400, "Mentors not found");
+    public Response updateMentorDetails(String token, Long id, MentorsDTO mentorsDTO) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Optional<MentorsModel> isMentorPresent = mentorsRepository.findById(id);
+            if (isMentorPresent.isPresent()) {
+                isMentorPresent.get().setEmployeeId(mentorsDTO.getEmployeeId());
+                isMentorPresent.get().setFirstName(mentorsDTO.getFirstName());
+                isMentorPresent.get().setLastName(mentorsDTO.getLastName());
+                isMentorPresent.get().setMentorDesc(mentorsDTO.getMentorDesc());
+                isMentorPresent.get().setMentorRole(mentorsDTO.getMentorRole());
+                isMentorPresent.get().setMentorsType(mentorsDTO.getMentorsType());
+                isMentorPresent.get().setEmail(mentorsDTO.getEmail());
+                isMentorPresent.get().setStatus(mentorsDTO.getStatus());
+                isMentorPresent.get().setExperienceYears(mentorsDTO.getExperienceYears());
+                isMentorPresent.get().setMobileNumber(mentorsDTO.getMobileNumber());
+                isMentorPresent.get().setStartDate(mentorsDTO.getStartDate());
+                isMentorPresent.get().setPreferredTime(mentorsDTO.getPreferredTime());
+                isMentorPresent.get().setCreatorUser(mentorsDTO.getCreatorUser());
+                isMentorPresent.get().setSupervisorId(mentorsDTO.getSupervisorId());
+                isMentorPresent.get().setUpdatedTimeStamp(LocalDateTime.now());
+                mentorsRepository.save(isMentorPresent.get());
+                String body = "Mentors Details Updated With Id is : " + isMentorPresent.get().getId();
+                String subject = "Mentors Details Updated Successfully ...";
+                mailService.send(isMentorPresent.get().getEmail(), body, subject);
+                return new Response(200, "Success", isMentorPresent.get());
+            } else {
+                throw new LMSException(400, "Mentors not found");
+            }
         }
-        return null;
+        throw new LMSException(400, "Token Wrong");
     }
 
     /*
@@ -118,17 +140,23 @@ public class MentorsService implements IMentorsService {
      * @Param :  token and id
      * */
     @Override
-    public MentorsModel deleteDetails(Long id) {
-        Optional<MentorsModel> isMentor = mentorsRepository.findById(id);
-        if (isMentor.isPresent()) {
-            mentorsRepository.delete(isMentor.get());
-            String body = "Mentor Details Deleted With Id is : " + isMentor.get().getId();
-            String subject = "Mentor Details Deleted Successfully ...";
-            mailService.send(isMentor.get().getEmail(), body, subject);
-            return isMentor.get();
-        } else {
-            throw new LMSException(400, "Not found with this id");
+    public Response deleteDetails(String token, Long id) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Optional<MentorsModel> isMentor = mentorsRepository.findById(id);
+            if (isMentor.isPresent()) {
+                mentorsRepository.delete(isMentor.get());
+                String body = "Mentor Details Deleted With Id is : " + isMentor.get().getId();
+                String subject = "Mentor Details Deleted Successfully ...";
+                mailService.send(isMentor.get().getEmail(), body, subject);
+                return new Response(200, "Success", isMentor.get());
+
+            } else {
+                throw new LMSException(400, "Not found with this id");
+            }
         }
+        throw new LMSException(400, "Token Wrong");
+
     }
 
     /*
@@ -137,13 +165,17 @@ public class MentorsService implements IMentorsService {
      * @Param :  token and bankDetailsDTO
      * */
     @Override
-    public MentorsModel getMentorsDetailsById(Long id) {
-        Optional<MentorsModel> isMentor = mentorsRepository.findById(id);
-        if (isMentor.isPresent()) {
-            return isMentor.get();
-        } else {
-            throw new LMSException(400, "Not found with this id");
+    public Response getMentorsDetailsById(String token, Long id) {
+        boolean isUserPresent = restTemplate.getForObject("http://localhost:8080/admin/validate/" + token, Boolean.class);
+        if (isUserPresent) {
+            Optional<MentorsModel> isMentor = mentorsRepository.findById(id);
+            if (isMentor.isPresent()) {
+                return new Response(200, "Success", isMentor);
+            } else {
+                throw new LMSException(400, "Not found with this id");
+            }
         }
+        throw new LMSException(400, "Token Wrong");
     }
 
     /*
